# Problem: many repeated GetParams calls in ante handler