--- a/app/ante/fee_market.go
+++ b/app/ante/fee_market.go
@@ -43,7 +43,6 @@ func NewGasWantedDecorator(
 }
 
 func (gwd GasWantedDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
-
 	blockHeight := big.NewInt(ctx.BlockHeight())
 	isLondon := gwd.ethCfg.IsLondon(blockHeight)