## Description

Solution:
- refactor the decorators and share the result
- move unprotected tx check to ValidateBasic (sigverify decorator could be removed in future PRs, see https://github.com/crypto-org-chain/ethermint/issues/355)

## Benchmark Results

The `BenchmarkERC20Transfer` in cronos repo.

```
goos: darwin
goarch: amd64
pkg: github.com/crypto-org-chain/cronos/v2/app
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                         │ /tmp/before.txt │           /tmp/after.txt           │
                         │     sec/op      │    sec/op     vs base              │
ERC20Transfer/memdb-12       654.9m ±   5%   612.9m ± 19%       ~ (p=0.132 n=6)
ERC20Transfer/leveldb-12     653.6m ± 150%   700.5m ± 18%       ~ (p=0.394 n=6)
ERC20Transfer/memiavl-12     627.6m ±  83%   604.1m ± 28%  -3.75% (p=0.041 n=6)
geomean                      645.3m          637.7m        -1.17%

                         │ /tmp/before.txt │           /tmp/after.txt            │
                         │      B/op       │     B/op      vs base               │
ERC20Transfer/memdb-12        203.8Mi ± 0%   176.4Mi ± 0%  -13.43% (p=0.002 n=6)
ERC20Transfer/leveldb-12      205.5Mi ± 0%   178.2Mi ± 0%  -13.30% (p=0.002 n=6)
ERC20Transfer/memiavl-12      198.4Mi ± 0%   171.0Mi ± 0%  -13.79% (p=0.002 n=6)
geomean                       202.5Mi        175.2Mi       -13.51%

                         │ /tmp/before.txt │           /tmp/after.txt           │
                         │    allocs/op    │  allocs/op   vs base               │
ERC20Transfer/memdb-12         3.284M ± 0%   2.802M ± 0%  -14.65% (p=0.002 n=6)
ERC20Transfer/leveldb-12       3.282M ± 0%   2.801M ± 0%  -14.65% (p=0.002 n=6)
ERC20Transfer/memiavl-12       3.228M ± 0%   2.747M ± 0%  -14.90% (p=0.002 n=6)
geomean                        3.264M        2.783M       -14.74%
```

<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

<!-- Add a description of the changes that this PR introduces and the files that
are the most critical to review.
-->

______

For contributor use:

- [ ] Targeted PR against correct branch (see [CONTRIBUTING.md](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#pr-targeting))
- [ ] Linked to Github issue with discussion and accepted design OR link to spec that describes this work.
- [ ] Code follows the [module structure standards](https://github.com/cosmos/cosmos-sdk/blob/master/docs/building-modules/structure.md).
- [ ] Wrote unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#testing)
- [ ] Updated relevant documentation (`docs/`) or specification (`x/<module>/spec/`)
- [ ] Added  relevant `godoc` [comments](https://blog.golang.org/godoc-documenting-go-code).
- [ ] Added a relevant changelog entry to the `Unreleased` section in `CHANGELOG.md`
- [ ] Re-reviewed `Files changed` in the Github PR explorer

______

For admin use:

- [ ] Added appropriate labels to PR (ex. `WIP`, `R4R`, `docs`, etc)
- [ ] Reviewers assigned
- [ ] Squashed all commits, uses message "Merge pull request #XYZ: [title]" ([coding standards](https://github.com/tendermint/coding/blob/master/README.md#merging-a-pr))